+++ lib/xmpp4r-simple_patch.rb
